Center button, remove unnecessary code and other fixes

+++ app/src/main/java/de/westnordost/streetcomplete/quests/LeaveNoteDialog.java
@@ -198,7 +198,10 @@ public class LeaveNoteDialog extends DialogFragment
 				imagePaths.add(photoFile.toString());
 			} else
 			{
-				photoFile.delete();
+				if (photoFile.exists())
+				{
+					photoFile.delete();
+				}
 			}
 		}
 	}
